Users can now log in and register. Log in is cached locally now, so no need to log in everytime.
Fixed views so they now work properly. Users can now register an
account and log in. Once logged in they will be brought to the search
page, this is the important part of the app. Will probably start that
tomorrow.

Updated UI a little bit too, may add screenshots of the views soon,
just for reference.